requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pandas
plotly
aiofiles
pytest
httpx
pdfkit
langchain
langchain-community
ydata-profiling
sweetviz
autoviz
lux
azure-ai-inference